+++ src/main/java/umc/th/juinjang/converter/limjang/LimjangUpdateConverter.java
@@ -1,19 +1,12 @@
 package umc.th.juinjang.converter.limjang;
 
-import static umc.th.juinjang.utils.LimjangUtil.checkExpectedSize;
-import static umc.th.juinjang.utils.LimjangUtil.makePriceList;
+import static umc.th.juinjang.service.LimjangService.LimjangPriceBridge.checkExpectedSize;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import umc.th.juinjang.apiPayload.code.status.ErrorStatus;
-import umc.th.juinjang.apiPayload.exception.handler.LimjangHandler;
-import umc.th.juinjang.model.dto.limjang.LimjangTotalListResponseDTO;
 import umc.th.juinjang.model.dto.limjang.LimjangUpdateRequestDTO;
 import umc.th.juinjang.model.entity.Limjang;
 import umc.th.juinjang.model.entity.enums.LimjangPriceType;
-import umc.th.juinjang.utils.LimjangUtil;
-import umc.th.juinjang.validation.annotation.VaildPriceListSize;
 
 public class LimjangUpdateConverter {
