+++ src/main/java/umc/th/juinjang/model/entity/LimjangPrice.java
@@ -6,14 +6,12 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
-import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import umc.th.juinjang.model.entity.common.BaseEntity;
-import umc.th.juinjang.utils.LimjangUtil;
 
 @Entity
 @Getter
